@@ -18,16 +18,16 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
         applicationId 'com.github.moko256.twitlatte'
         minSdkVersion 19
-        targetSdkVersion 27
-        versionCode 13
-        versionName '1.4.1'
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 16
+        versionName '1.5.2'
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
         Base64.Decoder e = Base64.decoder
 
