@@ -18,10 +18,11 @@ package com.github.moko256.twitlatte;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.preference.PreferenceFragmentCompat;
-import android.support.v7.preference.PreferenceScreen;
+
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.preference.PreferenceFragmentCompat;
+import androidx.preference.PreferenceScreen;
 
 /**
  * Created by moko256 on 2016/03/27.
@@ -33,7 +34,6 @@ public class SettingsActivity extends AppCompatActivity implements PreferenceFra
     @Override
     public void onCreate(Bundle savedInstanceState){
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_settings);
 
         ActionBar actionBar=getSupportActionBar();
         actionBar.setDisplayHomeAsUpEnabled(true);
@@ -58,7 +58,7 @@ public class SettingsActivity extends AppCompatActivity implements PreferenceFra
 
             getSupportFragmentManager()
                     .beginTransaction()
-                    .add(R.id.activity_settings_container,fragment)
+                    .add(android.R.id.content, fragment)
                     .commit();
         }
     }
