@@ -17,10 +17,11 @@
 package com.github.moko256.twitlatte.widget;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 /**
  * Created by moko256 on 2018/02/18.
@@ -28,7 +29,7 @@ import android.view.MotionEvent;
  * @author moko256
  */
 
-public class ViewPager extends android.support.v4.view.ViewPager {
+public class ViewPager extends androidx.viewpager.widget.ViewPager {
 
     public ViewPager(@NonNull Context context) {
         super(context);
