@@ -17,10 +17,10 @@
 package com.github.moko256.twitlatte
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.view.View
 import android.widget.ImageView
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 import java.util.*
 
 /**
