@@ -20,8 +20,6 @@ import android.content.Context;
 import android.database.Cursor;
 import android.net.Uri;
 import android.provider.MediaStore;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -35,6 +33,9 @@ import com.github.moko256.twitlatte.glide.GlideRequests;
 import java.util.ArrayList;
 import java.util.List;
 
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
 /**
  * Created by moko256 on 2017/03/11.
  *
@@ -45,8 +46,8 @@ public class AddedImagesAdapter extends RecyclerView.Adapter<RecyclerView.ViewHo
 
     private final Context context;
 
-    private ArrayList<Uri> images = new ArrayList<>();
     public int limit = 4;
+    private ArrayList<Uri> images = new ArrayList<>(4);
 
     public View.OnClickListener onAddButtonClickListener;
     public ImageAction onDeleteButtonListener;
@@ -54,6 +55,7 @@ public class AddedImagesAdapter extends RecyclerView.Adapter<RecyclerView.ViewHo
 
     public AddedImagesAdapter(Context context){
         this.context = context;
+        setHasStableIds(true);
     }
 
     @Override
