@@ -16,21 +16,14 @@
 
 package com.github.moko256.twitlatte;
 
-import android.arch.lifecycle.ViewModelProviders;
 import android.content.Context;
 import android.graphics.Point;
 import android.graphics.Rect;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.util.TypedValue;
 import android.view.Display;
 import android.view.Gravity;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.widget.Toast;
 
@@ -40,9 +33,16 @@ import com.github.moko256.twitlatte.viewmodel.ListViewModel;
 
 import java.util.Objects;
 
+import androidx.annotation.Nullable;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 import either.EitherKt;
+import io.reactivex.Single;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
+import io.reactivex.schedulers.Schedulers;
 import kotlin.Unit;
 import twitter4j.Paging;
 import twitter4j.ResponseList;
@@ -66,6 +66,8 @@ public abstract class BaseTweetListFragment extends BaseListFragment {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
         listViewModel = ViewModelProviders.of(this).get(ListViewModel.class);
         if (!listViewModel.getInitilized()) {
             listViewModel.statusIdsDatabase = new CachedIdListSQLiteOpenHelper(
@@ -88,17 +90,16 @@ public abstract class BaseTweetListFragment extends BaseListFragment {
             };
             listViewModel.start();
         }
-        super.onCreate(savedInstanceState);
     }
 
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view=super.onCreateView(inflater, container, savedInstanceState);
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
 
         int dp8 = Math.round(8f * getResources().getDisplayMetrics().density);
 
-        getRecyclerView().setPadding(dp8, 0, 0, 0);
-        getRecyclerView().addItemDecoration(new RecyclerView.ItemDecoration() {
+        recyclerView.setPadding(dp8, 0, 0, 0);
+        recyclerView.addItemDecoration(new RecyclerView.ItemDecoration() {
             @Override
             public void getItemOffsets(Rect outRect, View view, RecyclerView parent, RecyclerView.State state) {
                 outRect.right = dp8;
@@ -107,18 +108,161 @@ public abstract class BaseTweetListFragment extends BaseListFragment {
         });
 
         if (getActivity() instanceof GetRecyclerViewPool) {
-            getRecyclerView().setRecycledViewPool(((GetRecyclerViewPool) getActivity()).getTweetListViewPool());
+            recyclerView.setRecycledViewPool(((GetRecyclerViewPool) getActivity()).getTweetListViewPool());
         }
 
         adapter=new StatusesAdapter(getContext(), listViewModel.getList());
-        adapter.setOnLoadMoreClick(position -> listViewModel.loadOnGap(position));
-        setAdapter(adapter);
+        adapter.onLoadMoreClick = position -> listViewModel.loadOnGap(position);
+        adapter.onFavoriteClick = (position, id, hasFavorited) -> {
+            if (hasFavorited) {
+                Single
+                        .create(subscriber -> {
+                            try {
+                                Status newStatus = GlobalApplication.twitter.destroyFavorite(id);
+                                GlobalApplication.statusCache.add(newStatus, false);
+                                subscriber.onSuccess(newStatus);
+                            } catch (TwitterException e) {
+                                subscriber.tryOnError(e);
+                            }
+                        })
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(
+                                result -> {
+                                    if (adapter != null) {
+                                        adapter.notifyItemChanged(position);
+                                    }
+
+                                    notifyBySnackBar(
+                                            TwitterStringUtils.getDidActionStringRes(
+                                                    GlobalApplication.clientType,
+                                                    TwitterStringUtils.Action.UNLIKE
+                                            )
+                                    ).show();
+                                },
+                                throwable -> {
+                                    throwable.printStackTrace();
+                                    notifyErrorBySnackBar(throwable).show();
+                                    if (adapter != null) {
+                                        adapter.notifyItemChanged(position);
+                                    }
+                                }
+                        );
+            } else {
+                Single
+                        .create(subscriber -> {
+                            try {
+                                Status newStatus = GlobalApplication.twitter.createFavorite(id);
+                                GlobalApplication.statusCache.add(newStatus, false);
+                                subscriber.onSuccess(newStatus);
+                            } catch (TwitterException e) {
+                                subscriber.tryOnError(e);
+                            }
+                        })
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(
+                                result -> {
+                                    if (adapter != null) {
+                                        adapter.notifyItemChanged(position);
+                                    }
+
+                                    notifyBySnackBar(
+                                            TwitterStringUtils.getDidActionStringRes(
+                                                    GlobalApplication.clientType,
+                                                    TwitterStringUtils.Action.LIKE
+                                            )
+                                    ).show();
+                                },
+                                throwable -> {
+                                    throwable.printStackTrace();
+                                    notifyErrorBySnackBar(throwable).show();
+                                    if (adapter != null) {
+                                        adapter.notifyItemChanged(position);
+                                    }
+                                }
+                        );
+            }
+        };
+        adapter.onRepeatClick = (position, id, hasFavorited) -> {
+            if (hasFavorited) {
+                Single
+                        .create(subscriber -> {
+                            try {
+                                Status newStatus = GlobalApplication.twitter.unRetweetStatus(id);
+                                GlobalApplication.statusCache.add(newStatus, false);
+                                subscriber.onSuccess(newStatus);
+                            } catch (TwitterException e) {
+                                subscriber.tryOnError(e);
+                            }
+                        })
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(
+                                result -> {
+                                    if (adapter != null) {
+                                        adapter.notifyItemChanged(position);
+                                    }
+
+                                    notifyBySnackBar(
+                                            TwitterStringUtils.getDidActionStringRes(
+                                                    GlobalApplication.clientType,
+                                                    TwitterStringUtils.Action.UNREPEAT
+                                            )
+                                    ).show();
+                                },
+                                throwable -> {
+                                    throwable.printStackTrace();
+                                    notifyErrorBySnackBar(throwable).show();
+                                    if (adapter != null) {
+                                        adapter.notifyItemChanged(position);
+                                    }
+                                }
+                        );
+            } else {
+                Single
+                        .create(subscriber -> {
+                            try {
+                                Status newStatus = GlobalApplication.twitter.retweetStatus(id);
+                                GlobalApplication.statusCache.add(newStatus, false);
+                                subscriber.onSuccess(newStatus);
+                            } catch (TwitterException e) {
+                                subscriber.tryOnError(e);
+                            }
+                        })
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(
+                                result -> {
+                                    if (adapter != null) {
+                                        adapter.notifyItemChanged(position);
+                                    }
+
+                                    notifyBySnackBar(
+                                            TwitterStringUtils.getDidActionStringRes(
+                                                    GlobalApplication.clientType,
+                                                    TwitterStringUtils.Action.REPEAT
+                                            )
+                                    ).show();
+                                },
+                                throwable -> {
+                                    throwable.printStackTrace();
+                                    notifyErrorBySnackBar(throwable).show();
+                                    if (adapter != null) {
+                                        adapter.notifyItemChanged(position);
+                                    }
+                                }
+                        );
+            }
+        };
+
+        recyclerView.setAdapter(adapter);
         if (!isInitializedList()){
             adapter.notifyDataSetChanged();
         }
 
         LAST_SAVED_LIST_ID = listViewModel.getSeeingId();
-        getRecyclerView().getLayoutManager().scrollToPosition(listViewModel.getList().indexOf(LAST_SAVED_LIST_ID));
+        recyclerView.getLayoutManager().scrollToPosition(listViewModel.getList().indexOf(LAST_SAVED_LIST_ID));
 
         disposable = listViewModel.getListObserver()
                 .observeOn(AndroidSchedulers.mainThread())
@@ -154,8 +298,8 @@ public abstract class BaseTweetListFragment extends BaseListFragment {
                                         break;
 
                                     case INSERT_AT_GAP:
-                                        View startView = getRecyclerView().getLayoutManager().findViewByPosition(left.getPosition());
-                                        int offset = (startView == null) ? 0 : (startView.getTop() - getRecyclerView().getPaddingTop());
+                                        View startView = recyclerView.getLayoutManager().findViewByPosition(left.getPosition());
+                                        int offset = (startView == null) ? 0 : (startView.getTop() - recyclerView.getPaddingTop());
 
                                         boolean noGap = listViewModel.getList().get(left.getPosition() + left.getSize() - 1).equals(listViewModel.getList().get(left.getPosition() + 1));
 
@@ -165,7 +309,7 @@ public abstract class BaseTweetListFragment extends BaseListFragment {
                                             adapter.notifyItemChanged(left.getPosition());
                                         }
 
-                                        RecyclerView.LayoutManager layoutManager = getRecyclerView().getLayoutManager();
+                                        RecyclerView.LayoutManager layoutManager = recyclerView.getLayoutManager();
                                         if (layoutManager instanceof LinearLayoutManager) {
                                             ((LinearLayoutManager) layoutManager).scrollToPositionWithOffset(left.getPosition() + left.getSize(), offset);
                                             adapter.notifyItemRangeInserted(left.getPosition(), left.getSize());
@@ -177,32 +321,31 @@ public abstract class BaseTweetListFragment extends BaseListFragment {
                                 return Unit.INSTANCE;
                             },
                             right -> {
-                                getSnackBar(TwitterStringUtils.convertErrorToText(right)).show();
+                                notifyErrorBySnackBar(right).show();
                                 return Unit.INSTANCE;
                             }
                     );
-                    if (getSwipeRefreshLayout().isRefreshing()){
+                    if (swipeRefreshLayout.isRefreshing()){
                         setRefreshing(false);
                     }
                 });
-
-        return view;
     }
 
     @Override
     public void onDestroyView() {
         disposable.dispose();
-        RecyclerView.LayoutManager layoutManager = getRecyclerView().getLayoutManager();
+        RecyclerView.LayoutManager layoutManager = recyclerView.getLayoutManager();
         int position = getFirstVisibleItemPosition(layoutManager);
         listViewModel.removeOldCache(position);
-        super.onDestroyView();
+        recyclerView.swapAdapter(null, true);
         adapter=null;
+        super.onDestroyView();
     }
 
     @Override
     public void onStop() {
         super.onStop();
-        int position = getFirstVisibleItemPosition(getRecyclerView().getLayoutManager());
+        int position = getFirstVisibleItemPosition(recyclerView.getLayoutManager());
         if (position >= 0) {
             long id = listViewModel.getList().get(
                     position
