@@ -18,17 +18,14 @@ package com.github.moko256.twitlatte;
 
 import android.graphics.Rect;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
-
-import com.github.moko256.twitlatte.text.TwitterStringUtils;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 import io.reactivex.Observable;
 import io.reactivex.Single;
 import io.reactivex.android.schedulers.AndroidSchedulers;
@@ -53,16 +50,27 @@ public abstract class BaseUsersFragment extends BaseListFragment {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
         list=new ArrayList<>();
         disposable = new CompositeDisposable();
-        super.onCreate(savedInstanceState);
+
+        if (savedInstanceState != null) {
+            long[] l = savedInstanceState.getLongArray("list");
+            if(l != null){
+                for (long id : l) {
+                    list.add(id);
+                }
+            }
+            next_cursor=savedInstanceState.getLong("next_cursor",-1);
+        }
     }
 
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view=super.onCreateView(inflater, container, savedInstanceState);
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
 
-        getRecyclerView().addItemDecoration(new RecyclerView.ItemDecoration() {
+        recyclerView.addItemDecoration(new RecyclerView.ItemDecoration() {
             @Override
             public void getItemOffsets(Rect outRect, View view, RecyclerView parent, RecyclerView.State state) {
                 super.getItemOffsets(outRect, view, parent, state);
@@ -73,29 +81,13 @@ public abstract class BaseUsersFragment extends BaseListFragment {
         });
 
         if (getActivity() instanceof BaseTweetListFragment.GetRecyclerViewPool) {
-            getRecyclerView().setRecycledViewPool(((GetRecyclerViewPool) getActivity()).getUserListViewPool());
+            recyclerView.setRecycledViewPool(((GetRecyclerViewPool) getActivity()).getUserListViewPool());
         }
 
         adapter=new UsersAdapter(getContext(), list);
-        setAdapter(adapter);
+        recyclerView.setAdapter(adapter);
         if(!isInitializedList()){
             adapter.notifyDataSetChanged();
-        }
-
-        return view;
-    }
-
-    @Override
-    public void onRestoreInstanceState(Bundle savedInstanceState) {
-        super.onRestoreInstanceState(savedInstanceState);
-        if (savedInstanceState != null) {
-            long[] l = savedInstanceState.getLongArray("list");
-            if(l != null){
-                for (long id : l) {
-                    list.add(id);
-                }
-            }
-            next_cursor=savedInstanceState.getLong("next_cursor",-1);
         }
     }
 
@@ -113,8 +105,9 @@ public abstract class BaseUsersFragment extends BaseListFragment {
 
     @Override
     public void onDestroyView() {
-        super.onDestroyView();
+        recyclerView.swapAdapter(null, true);
         adapter=null;
+        super.onDestroyView();
     }
 
     @Override
@@ -149,7 +142,7 @@ public abstract class BaseUsersFragment extends BaseListFragment {
                                 },
                                 e -> {
                                     e.printStackTrace();
-                                    getSnackBar(TwitterStringUtils.convertErrorToText(e)).show();
+                                    notifyErrorBySnackBar(e).show();
                                 }
                         )
         );
@@ -184,7 +177,7 @@ public abstract class BaseUsersFragment extends BaseListFragment {
                                 },
                                 e -> {
                                     e.printStackTrace();
-                                    getSnackBar(TwitterStringUtils.convertErrorToText(e)).show();
+                                    notifyErrorBySnackBar(e).show();
                                 }
                         )
         );
