@@ -17,19 +17,16 @@
 package com.github.moko256.twitlatte;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
 
 import com.github.moko256.mastodon.MTException;
 import com.github.moko256.mastodon.MTRangeConverter;
 import com.github.moko256.mastodon.MTResponseList;
 import com.github.moko256.mastodon.MastodonTwitterImpl;
-import com.github.moko256.twitlatte.entity.Type;
+import com.github.moko256.twitlatte.entity.ClientType;
 import com.sys1yagi.mastodon4j.api.exception.Mastodon4jRequestException;
 import com.sys1yagi.mastodon4j.api.method.Accounts;
 
+import androidx.annotation.Nullable;
 import twitter4j.Paging;
 import twitter4j.ResponseList;
 import twitter4j.Status;
@@ -49,14 +46,15 @@ public class MediaTimelineFragment extends BaseTweetListFragment implements Tool
         if (userId == -1){
             userId = getArguments().getLong("userId", -1);
         }
+
         super.onCreate(savedInstanceState);
     }
 
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = super.onCreateView(inflater, container, savedInstanceState);
-        adapter.setShouldShowMediaOnly(true);
-        return view;
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+
+        adapter.shouldShowMediaOnly = true;
     }
 
     public static MediaTimelineFragment newInstance(long userId){
@@ -69,7 +67,7 @@ public class MediaTimelineFragment extends BaseTweetListFragment implements Tool
 
     @Override
     protected ResponseList<Status> getResponseList(Paging paging) throws TwitterException {
-        if (GlobalApplication.clientType == Type.MASTODON){
+        if (GlobalApplication.clientType == ClientType.MASTODON){
             Accounts accounts = new Accounts(((MastodonTwitterImpl) GlobalApplication.twitter).client);
             try {
                 return MTResponseList.convert(accounts.getStatuses(userId, true, false, false, MTRangeConverter.convert(paging)).execute());
