@@ -16,38 +16,40 @@
 
 package com.github.moko256.twitlatte;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
 import android.text.method.LinkMovementMethod;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.FrameLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.github.moko256.twitlatte.cacheMap.StatusCacheMap;
+import com.github.moko256.twitlatte.entity.Post;
+import com.github.moko256.twitlatte.entity.Repeat;
+import com.github.moko256.twitlatte.entity.Status;
+import com.github.moko256.twitlatte.entity.User;
+import com.github.moko256.twitlatte.glide.GlideApp;
 import com.github.moko256.twitlatte.intent.AppCustomTabsKt;
 import com.github.moko256.twitlatte.model.base.PostTweetModel;
 import com.github.moko256.twitlatte.model.impl.PostTweetModelCreator;
-import com.github.moko256.twitlatte.text.MTHtmlParser;
 import com.github.moko256.twitlatte.text.TwitterStringUtils;
 
 import java.text.DateFormat;
 
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityOptionsCompat;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import io.reactivex.Single;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.CompositeDisposable;
 import io.reactivex.schedulers.Schedulers;
-import twitter4j.Status;
 import twitter4j.TwitterException;
-import twitter4j.User;
 
 import static android.view.View.GONE;
 import static android.view.View.VISIBLE;
@@ -62,11 +64,11 @@ public class ShowTweetActivity extends AppCompatActivity {
 
     private CompositeDisposable disposables;
     private long statusId;
+    private String shareUrl = "";
 
-    private StatusView statusView;
+    private StatusViewBinder statusViewBinder;
 
     private TextView tweetIsReply;
-    private FrameLayout statusViewFrame;
     private TextView timestampText;
     private TextView viaText;
     private EditText replyText;
@@ -90,7 +92,8 @@ public class ShowTweetActivity extends AppCompatActivity {
             }
 
             tweetIsReply = findViewById(R.id.tweet_show_is_reply_text);
-            statusViewFrame = findViewById(R.id.tweet_show_tweet);
+            ViewGroup statusViewFrame = findViewById(R.id.tweet_show_tweet);
+            statusViewBinder = new StatusViewBinder(GlideApp.with(this), statusViewFrame);
             timestampText = findViewById(R.id.tweet_show_timestamp);
             viaText = findViewById(R.id.tweet_show_via);
             replyText= findViewById(R.id.tweet_show_tweet_reply_text);
@@ -118,28 +121,7 @@ public class ShowTweetActivity extends AppCompatActivity {
                                     })
             ));
 
-            Status status = GlobalApplication.statusCache.get(statusId);
-            if (status == null){
-                disposables.add(
-                        updateStatus()
-                                .subscribeOn(Schedulers.io())
-                                .observeOn(AndroidSchedulers.mainThread())
-                                .subscribe(
-                                        result->{
-                                            if (result == null) {
-                                                finish();
-                                                return;
-                                            }
-                                            updateView(result);
-                                        },
-                                        e->{
-                                            e.printStackTrace();
-                                            finish();
-                                        })
-                );
-            } else {
-                updateView(status);
-            }
+            prepareStatus();
         }
     }
 
@@ -148,10 +130,10 @@ public class ShowTweetActivity extends AppCompatActivity {
         disposables.dispose();
         super.onDestroy();
         disposables=null;
-        if (statusView != null){
-            statusView.setStatus(null);
+        if (statusViewBinder != null){
+            statusViewBinder.clear();
         }
-        statusView=null;
+        statusViewBinder=null;
     }
 
 
@@ -161,17 +143,13 @@ public class ShowTweetActivity extends AppCompatActivity {
         return super.onCreateOptionsMenu(menu);
     }
 
-    private String getShareUrl() {
-        return ((StatusCacheMap.CachedStatus) GlobalApplication.statusCache.get(statusId)).getRemoteUrl();
-    }
-
-        @Override
+    @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()){
             case R.id.action_quote:
                 startActivity(PostActivity.getIntent(
                         this,
-                        getShareUrl() + " "
+                        shareUrl + " "
                 ));
                 break;
             case R.id.action_share:
@@ -179,11 +157,11 @@ public class ShowTweetActivity extends AppCompatActivity {
                         new Intent()
                                 .setAction(Intent.ACTION_SEND)
                                 .setType("text/plain")
-                                .putExtra(Intent.EXTRA_TEXT, getShareUrl()),
+                                .putExtra(Intent.EXTRA_TEXT, shareUrl),
                         getString(R.string.share)));
                 break;
             case R.id.action_open_in_browser:
-                AppCustomTabsKt.launchChromeCustomTabs(this, getShareUrl());
+                AppCustomTabsKt.launchChromeCustomTabs(this, shareUrl);
                 break;
             }
         return super.onOptionsItemSelected(item);
@@ -199,21 +177,49 @@ public class ShowTweetActivity extends AppCompatActivity {
         return new Intent(context, ShowTweetActivity.class).putExtra("statusId", statusId);
     }
 
-    private Single<Status> updateStatus(){
+    private void prepareStatus() {
+        Post<Repeat, Status, User> status = GlobalApplication.postCache.getPost(statusId);
+
+        if (status != null) {
+            updateView(status);
+        } else {
+            disposables.add(
+                    updateStatus()
+                            .subscribeOn(Schedulers.io())
+                            .observeOn(AndroidSchedulers.mainThread())
+                            .subscribe(
+                                    result->{
+                                        if (result == null) {
+                                            finish();
+                                            return;
+                                        }
+                                        updateView(result);
+                                    },
+                                    e->{
+                                        e.printStackTrace();
+                                        finish();
+                                    })
+            );
+        }
+    }
+
+    private Single<Post<Repeat, Status, User>> updateStatus(){
         return Single.create(
                 subscriber -> {
                     try {
-                        Status status = GlobalApplication.twitter.showStatus(statusId);
-                        GlobalApplication.statusCache.add(status, false);
-                        subscriber.onSuccess(GlobalApplication.statusCache.get(statusId));
+                        twitter4j.Status result = GlobalApplication.twitter.showStatus(statusId);
+                        GlobalApplication.statusCache.add(result, false);
+                        subscriber.onSuccess(GlobalApplication.postCache.getPost(statusId));
                     } catch (TwitterException e) {
                         subscriber.tryOnError(e);
                     }
                 });
     }
 
-    private void updateView(Status item){
-        long replyTweetId = item.getInReplyToStatusId();
+    @SuppressLint("CheckResult")
+    private void updateView(Post<Repeat, Status, User> item){
+        shareUrl = item.getStatus().getUrl();
+        long replyTweetId = item.getStatus().getInReplyToStatusId();
         if (replyTweetId != -1){
             tweetIsReply.setVisibility(VISIBLE);
             tweetIsReply.setOnClickListener(v -> startActivity(getIntent(this, replyTweetId)));
@@ -221,37 +227,224 @@ public class ShowTweetActivity extends AppCompatActivity {
             tweetIsReply.setVisibility(GONE);
         }
 
-        statusView = new StatusView(this);
-        statusView.setStatus(item);
-        statusView.setOnClickListener(null);
-        ViewGroup cview = (ViewGroup) statusView.getChildAt(0);
-        ViewGroup sview = (ViewGroup) cview.getChildAt(0);
-        cview.removeView(sview);
-        statusViewFrame.removeAllViews();
-        statusViewFrame.addView(sview);
+        statusViewBinder.getUserImage().setOnClickListener(v -> {
+            ActivityOptionsCompat animation = ActivityOptionsCompat
+                    .makeSceneTransitionAnimation(
+                            this,
+                            v,
+                            "icon_image"
+                    );
+            startActivity(
+                    ShowUserActivity.getIntent(this, item.getUser().getId()),
+                    animation.toBundle()
+            );
+        });
+
+        statusViewBinder.getQuoteTweetLayout().setOnClickListener(v -> startActivity(
+                ShowTweetActivity.getIntent(this, item.getQuotedRepeatingStatus().getId())
+        ));
+
+        statusViewBinder.getLikeButton().setOnCheckedChangeListener((compoundButton, b) -> {
+            if (b && (!item.getStatus().isFavorited())) {
+                Single
+                        .create(subscriber -> {
+                            try {
+                                twitter4j.Status newStatus = GlobalApplication.twitter.createFavorite(item.getStatus().getId());
+                                GlobalApplication.statusCache.add(newStatus, false);
+                                subscriber.onSuccess(newStatus);
+                            } catch (TwitterException e) {
+                                subscriber.tryOnError(e);
+                            }
+                        })
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(
+                                result -> {
+                                    if (statusViewBinder != null) {
+                                        prepareStatus();
+                                    }
+
+                                    Toast.makeText(
+                                            this,
+                                            TwitterStringUtils.getDidActionStringRes(
+                                                    GlobalApplication.clientType,
+                                                    TwitterStringUtils.Action.LIKE
+                                            ),
+                                            Toast.LENGTH_SHORT
+                                    ).show();
+                                },
+                                throwable -> {
+                                    throwable.printStackTrace();
+                                    Toast.makeText(
+                                            this,
+                                            TwitterStringUtils.convertErrorToText(throwable),
+                                            Toast.LENGTH_LONG
+                                    ).show();
+                                }
+                        );
+            } else if ((!b) && item.getStatus().isFavorited()) {
+                Single
+                        .create(subscriber -> {
+                            try {
+                                twitter4j.Status newStatus = GlobalApplication.twitter.destroyFavorite(item.getStatus().getId());
+                                GlobalApplication.statusCache.add(newStatus, false);
+                                subscriber.onSuccess(newStatus);
+                            } catch (TwitterException e) {
+                                subscriber.tryOnError(e);
+                            }
+                        })
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(
+                                result -> {
+                                    if (statusViewBinder != null) {
+                                        prepareStatus();
+                                    }
+
+                                    Toast.makeText(
+                                            this,
+                                            TwitterStringUtils.getDidActionStringRes(
+                                                    GlobalApplication.clientType,
+                                                    TwitterStringUtils.Action.UNLIKE
+                                            ),
+                                            Toast.LENGTH_SHORT
+                                    ).show();
+                                },
+                                throwable -> {
+                                    throwable.printStackTrace();
+                                    Toast.makeText(
+                                            this,
+                                            TwitterStringUtils.convertErrorToText(throwable),
+                                            Toast.LENGTH_LONG
+                                    ).show();
+                                }
+                        );
+            }
+        });
+
+        statusViewBinder.getRetweetButton().setOnCheckedChangeListener((compoundButton, b) -> {
+            if (b && (!item.getStatus().isRepeated())) {
+                Single
+                        .create(subscriber -> {
+                            try {
+                                twitter4j.Status newStatus = GlobalApplication.twitter.retweetStatus(item.getStatus().getId());
+                                GlobalApplication.statusCache.add(newStatus, false);
+                                subscriber.onSuccess(newStatus);
+                            } catch (TwitterException e) {
+                                subscriber.tryOnError(e);
+                            }
+                        })
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(
+                                result -> {
+                                    if (statusViewBinder != null) {
+                                        prepareStatus();
+                                    }
+
+                                    Toast.makeText(
+                                            this,
+                                            TwitterStringUtils.getDidActionStringRes(
+                                                    GlobalApplication.clientType,
+                                                    TwitterStringUtils.Action.REPEAT
+                                            ),
+                                            Toast.LENGTH_SHORT
+                                    ).show();
+                                },
+                                throwable -> {
+                                    throwable.printStackTrace();
+                                    Toast.makeText(
+                                            this,
+                                            TwitterStringUtils.convertErrorToText(throwable),
+                                            Toast.LENGTH_LONG
+                                    ).show();
+                                }
+                        );
+            } else if ((!b) && item.getStatus().isRepeated()) {
+                Single
+                        .create(subscriber -> {
+                            try {
+                                twitter4j.Status newStatus = GlobalApplication.twitter.unRetweetStatus(item.getStatus().getId());
+                                GlobalApplication.statusCache.add(newStatus, false);
+                                subscriber.onSuccess(newStatus);
+                            } catch (TwitterException e) {
+                                subscriber.tryOnError(e);
+                            }
+                        })
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(
+                                result -> {
+                                    if (statusViewBinder != null) {
+                                        prepareStatus();
+                                    }
+
+                                    Toast.makeText(
+                                            this,
+                                            TwitterStringUtils.getDidActionStringRes(
+                                                    GlobalApplication.clientType,
+                                                    TwitterStringUtils.Action.UNREPEAT
+                                            ),
+                                            Toast.LENGTH_SHORT
+                                    ).show();
+                                },
+                                throwable -> {
+                                    throwable.printStackTrace();
+                                    Toast.makeText(
+                                            this,
+                                            TwitterStringUtils.convertErrorToText(throwable),
+                                            Toast.LENGTH_LONG
+                                    ).show();
+                                }
+                        );
+            }
+        });
+
+        statusViewBinder.getReplyButton().setOnClickListener(
+                v -> startActivity(PostActivity.getIntent(
+                        this,
+                        item.getStatus().getId(),
+                        TwitterStringUtils.convertToReplyTopString(
+                                GlobalApplication.userCache.get(GlobalApplication.userId).getScreenName(),
+                                item.getUser().getScreenName(),
+                                item.getStatus().getMentions()
+                        ).toString()
+                ))
+        );
+
+        statusViewBinder.setStatus(
+                item.getUser(),
+                item.getStatus(),
+                item.getQuotedRepeatingUser(),
+                item.getQuotedRepeatingStatus()
+        );
 
         timestampText.setText(
                 DateFormat.getDateTimeInstance(DateFormat.FULL, DateFormat.FULL)
-                        .format(item.getCreatedAt())
+                        .format(item.getStatus().getCreatedAt())
         );
 
-        viaText.setText(MTHtmlParser.INSTANCE.convertToEntities("via:"+item.getSource(), TwitterStringUtils.linkParserListener(this)));
-        viaText.setMovementMethod(new LinkMovementMethod());
+        if (item.getStatus().getSourceName() != null) {
+            viaText.setText(TwitterStringUtils.appendLinkAtViaText(this, item.getStatus().getSourceName(), item.getStatus().getSourceWebsite()));
+            viaText.setMovementMethod(new LinkMovementMethod());
+        } else {
+            viaText.setVisibility(GONE);
+        }
 
-        resetReplyText(item);
+        resetReplyText(item.getUser(), item.getStatus());
 
         replyButton.setOnClickListener(v -> {
             replyButton.setEnabled(false);
             PostTweetModel model = PostTweetModelCreator.getInstance(GlobalApplication.twitter, getContentResolver());
             model.setTweetText(replyText.getText().toString());
-            model.setInReplyToStatusId(item.getId());
+            model.setInReplyToStatusId(item.getStatus().getId());
             disposables.add(
                     model.postTweet()
                             .subscribeOn(Schedulers.io())
                             .observeOn(AndroidSchedulers.mainThread())
                             .subscribe(
                                     () -> {
-                                        resetReplyText(item);
+                                        resetReplyText(item.getUser(), item.getStatus());
                                         replyButton.setEnabled(true);
                                         Toast.makeText(ShowTweetActivity.this,R.string.did_post,Toast.LENGTH_SHORT).show();
                                     },
@@ -265,14 +458,13 @@ public class ShowTweetActivity extends AppCompatActivity {
         });
     }
 
-    private void resetReplyText(Status status){
+    private void resetReplyText(User postedUser, Status status){
         User user = GlobalApplication.userCache.get(GlobalApplication.userId);
 
         replyText.setText(TwitterStringUtils.convertToReplyTopString(
                 user != null ? user.getScreenName() : GlobalApplication.accessToken.getScreenName(),
-                status.getUser().getScreenName(),
-                status.getUserMentionEntities()
+                postedUser.getScreenName(),
+                status.getMentions()
         ));
     }
-
 }
