@@ -19,41 +19,31 @@ package com.github.moko256.twitlatte.text;
 import android.content.Context;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.graphics.drawable.DrawableCompat;
-import android.support.v7.content.res.AppCompatResources;
 import android.text.SpannableStringBuilder;
 import android.text.Spanned;
 import android.text.TextUtils;
-import android.text.style.ClickableSpan;
 import android.text.style.ImageSpan;
 import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.github.moko256.mastodon.MTException;
 import com.github.moko256.twitlatte.GlobalApplication;
 import com.github.moko256.twitlatte.R;
 import com.github.moko256.twitlatte.SearchResultActivity;
 import com.github.moko256.twitlatte.ShowUserActivity;
-import com.github.moko256.twitlatte.entity.Type;
+import com.github.moko256.twitlatte.entity.ClientType;
 import com.github.moko256.twitlatte.intent.AppCustomTabsKt;
+import com.github.moko256.twitlatte.text.link.entity.Link;
+import com.github.moko256.twitlatte.text.style.ClickableNoLineSpan;
 import com.sys1yagi.mastodon4j.api.exception.Mastodon4jRequestException;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import kotlin.jvm.functions.Function1;
-import twitter4j.HashtagEntity;
-import twitter4j.Status;
-import twitter4j.SymbolEntity;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
+import androidx.appcompat.content.res.AppCompatResources;
+import androidx.core.content.ContextCompat;
+import androidx.core.graphics.drawable.DrawableCompat;
 import twitter4j.TwitterException;
-import twitter4j.URLEntity;
-import twitter4j.User;
-import twitter4j.UserMentionEntity;
 
 /**
  * Created by moko256 on 2016/08/06.
@@ -64,51 +54,49 @@ import twitter4j.UserMentionEntity;
 public class TwitterStringUtils {
 
     @NonNull
-    public static String plusAtMark(String... strings){
+    public static StringBuilder plusAtMark(String... strings){
         StringBuilder stringBuilder = new StringBuilder();
         for (String string: strings) {
             stringBuilder.append("@").append(string);
         }
-        return stringBuilder.toString();
+        return stringBuilder;
     }
 
-    public static String convertToSIUnitString(int num){
+    public static CharSequence convertToSIUnitString(int num){
         if (num == 0) return "0";
-        boolean isNegative = (num < 0);
-        String sign;
-        if (isNegative) {
+        StringBuilder builder = new StringBuilder(5);
+        if (num < 0) {
             num *= -1;
-            sign = "-";
-        } else {
-            sign = "";
+            builder.append("-");
         }
 
         float k = num / 1000;
-        if (k < 1) return sign + String.valueOf(num);
+        if (k < 1) return builder.append(String.valueOf(num));
 
         float m = k / 1000;
-        if (m < 1) return sign + String.valueOf(Math.round(k)) + "K";
+        if (m < 1) return builder.append(String.valueOf(Math.round(k))).append("K");
 
         float g = m / 1000;
-        if (g < 1) return sign + String.valueOf(Math.round(m)) + "M";
+        if (g < 1) return builder.append(String.valueOf(Math.round(m))).append("M");
 
-        return sign + String.valueOf(Math.round(g)) + "G";
+        return builder.append(String.valueOf(Math.round(g))).append("G");
     }
 
     @NonNull
     public static CharSequence convertToReplyTopString(@NonNull String userScreenName,
                                                        @NonNull String replyToScreenName,
-                                                       @NonNull UserMentionEntity[] users){
+                                                       @Nullable String[] users){
         StringBuilder userIdsStr = new StringBuilder();
 
         if (!userScreenName.equals(replyToScreenName)) {
             userIdsStr.append("@").append(replyToScreenName).append(" ");
         }
 
-        for (UserMentionEntity user : users) {
-            String screenName = user.getScreenName();
-            if (!(screenName.equals(userScreenName) || screenName.equals(replyToScreenName))) {
-                userIdsStr.append("@").append(screenName).append(" ");
+        if (users != null) {
+            for (String screenName : users) {
+                if (!(screenName.equals(userScreenName) || screenName.equals(replyToScreenName))) {
+                    userIdsStr.append("@").append(screenName).append(" ");
+                }
             }
         }
         return userIdsStr;
@@ -125,191 +113,108 @@ public class TwitterStringUtils {
         }
     }
 
-    public static CharSequence getLinkedSequence(Context context, Status item){
-
-        String tweet = item.getText();
-
-        if (tweet == null) {
-            return "";
+    public static CharSequence getLinkedSequence(Context context, String text, Link[] links){
+        if (links == null) {
+            return text;
         }
 
-        if (GlobalApplication.clientType == Type.MASTODON){
-            CharSequence html = MTHtmlParser.INSTANCE.convertToEntities(tweet, linkParserListener(context));
-            int length = html.length();
-            // If post has media only, context of post from Mastodon is "."
-            if (length == 1
-                    && item.getMediaEntities().length > 0
-                    && html.charAt(0) == ".".charAt(0)
-                    ){
-                return "";
-            } else {
-                return html;
-            }
-        } else {
+        SpannableStringBuilder spannableStringBuilder = new SpannableStringBuilder(text);
 
-            SpannableStringBuilder spannableStringBuilder = new SpannableStringBuilder(tweet);
+        for (Link link : links) {
+            Object span;
 
-            try {
-                for (SymbolEntity symbolEntity : item.getSymbolEntities()) {
-                    spannableStringBuilder.setSpan(new ClickableSpan() {
-                        @Override
-                        public void onClick(View view) {
-                            context.startActivity(SearchResultActivity.getIntent(context, "$" + symbolEntity.getText()));
-                        }
-                    }, tweet.offsetByCodePoints(0, symbolEntity.getStart()), tweet.offsetByCodePoints(0, symbolEntity.getEnd()), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-                }
-
-                for (HashtagEntity hashtagEntity : item.getHashtagEntities()) {
-                    spannableStringBuilder.setSpan(new ClickableSpan() {
-                        @Override
-                        public void onClick(View view) {
-                            context.startActivity(SearchResultActivity.getIntent(context, "#" + hashtagEntity.getText())
-                            );
-                        }
-                    }, tweet.offsetByCodePoints(0, hashtagEntity.getStart()), tweet.offsetByCodePoints(0, hashtagEntity.getEnd()), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-                }
-
-                for (UserMentionEntity userMentionEntity : item.getUserMentionEntities()) {
-                    spannableStringBuilder.setSpan(new ClickableSpan() {
-                        @Override
-                        public void onClick(View view) {
-                            context.startActivity(
-                                    ShowUserActivity.getIntent(context, userMentionEntity.getScreenName())
-                            );
-                        }
-                    }, tweet.offsetByCodePoints(0, userMentionEntity.getStart()), tweet.offsetByCodePoints(0, userMentionEntity.getEnd()), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-                }
-
-                boolean hasMedia = item.getMediaEntities().length > 0;
-                List<URLEntity> urlEntities = new ArrayList<>(item.getURLEntities().length + (hasMedia ? 1 : 0));
-                urlEntities.addAll(Arrays.asList(item.getURLEntities()));
-                if (hasMedia) {
-                    urlEntities.add(item.getMediaEntities()[0]);
-                }
-
-                int tweetLength = tweet.codePointCount(0, tweet.length());
-                int sp = 0;
-
-                for (URLEntity entity : urlEntities) {
-                    String url = entity.getURL();
-                    String displayUrl = entity.getDisplayURL();
-
-                    int urlLength = url.codePointCount(0, url.length());
-                    int displayUrlLength = displayUrl.codePointCount(0, displayUrl.length());
-                    if (entity.getStart() <= tweetLength && entity.getEnd() <= tweetLength) {
-                        int dusp = displayUrlLength - urlLength;
-                        spannableStringBuilder.replace(tweet.offsetByCodePoints(0, entity.getStart()) + sp, tweet.offsetByCodePoints(0, entity.getEnd()) + sp, displayUrl);
-                        spannableStringBuilder.setSpan(new ClickableSpan() {
+            final Uri uri = Uri.parse(link.getUrl());
+            if (uri.getScheme().equals("twitlatte")) {
+                switch (uri.getHost()) {
+                    case "symbol":
+                        span = new ClickableNoLineSpan() {
                             @Override
                             public void onClick(View view) {
-                                AppCustomTabsKt.launchChromeCustomTabs(context, entity.getExpandedURL());
+                                context.startActivity(
+                                        SearchResultActivity.getIntent(context, "$" + uri.getLastPathSegment())
+                                );
+
                             }
-                        }, tweet.offsetByCodePoints(0, entity.getStart()) + sp, tweet.offsetByCodePoints(0, entity.getEnd()) + sp + dusp, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-
-                        sp += dusp;
-                    }
-                }
-            } catch (IndexOutOfBoundsException e) {
-                e.printStackTrace();
-                Toast.makeText(context, R.string.error_occurred, Toast.LENGTH_SHORT).show();
-            }
-
-            return spannableStringBuilder;
-        }
-    }
-
-    public static CharSequence getProfileLinkedSequence(Context context, User user){
-
-        String description = user.getDescription();
-
-        if (GlobalApplication.clientType == Type.MASTODON){
-            return MTHtmlParser.INSTANCE.convertToEntities(description, linkParserListener(context));
-        } else {
-            URLEntity[] urlEntities = user.getDescriptionURLEntities();
-
-            if (urlEntities == null || urlEntities.length <= 0 || TextUtils.isEmpty(urlEntities[0].getURL()))
-                return description;
-
-            SpannableStringBuilder spannableStringBuilder = new SpannableStringBuilder(description);
-            int tweetLength = description.codePointCount(0, description.length());
-            int sp = 0;
-
-            for (URLEntity entity : urlEntities) {
-                String url = entity.getURL();
-                String expandedUrl = entity.getDisplayURL();
-
-                int urlLength = url.codePointCount(0, url.length());
-                int displayUrlLength = expandedUrl.codePointCount(0, expandedUrl.length());
-                if (entity.getStart() <= tweetLength && entity.getEnd() <= tweetLength) {
-                    int dusp = displayUrlLength - urlLength;
-                    spannableStringBuilder.replace(description.offsetByCodePoints(0, entity.getStart()) + sp, description.offsetByCodePoints(0, entity.getEnd()) + sp, expandedUrl);
-                    spannableStringBuilder.setSpan(new ClickableSpan() {
-                        @Override
-                        public void onClick(View view) {
-                            AppCustomTabsKt.launchChromeCustomTabs(context, entity.getExpandedURL());
-                        }
-                    }, description.offsetByCodePoints(0, entity.getStart()) + sp, description.offsetByCodePoints(0, entity.getEnd()) + sp + dusp, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-
-                    sp += dusp;
-                }
-            }
-
-            return spannableStringBuilder;
-        }
-    }
-
-    public static Function1<String, Object> linkParserListener(Context context) {
-         return link -> {
-            final Uri uri = Uri.parse(link);
-            if (uri.getScheme().equals("twitlatte")) {
-                if (uri.getHost().equals("user")) {
-                    return new ClickableSpan() {
-                        @Override
-                        public void onClick(View view) {
-                            context.startActivity(
-                                    ShowUserActivity.getIntent(context, uri.getLastPathSegment())
-                            );
-                        }
-                    };
-                } else {
-                    return new ClickableSpan() {
-                        @Override
-                        public void onClick(View view) {
-                            context.startActivity(SearchResultActivity.getIntent(context, uri.getLastPathSegment()));
-                        }
-                    };
+                        };
+                        break;
+                    case "hashtag":
+                        span = new ClickableNoLineSpan() {
+                            @Override
+                            public void onClick(View view) {
+                                context.startActivity(
+                                        SearchResultActivity.getIntent(context, "#" + uri.getLastPathSegment())
+                                );
+                            }
+                        };
+                        break;
+                    case "user":
+                        span = new ClickableNoLineSpan() {
+                            @Override
+                            public void onClick(View view) {
+                                context.startActivity(
+                                        ShowUserActivity.getIntent(context, uri.getLastPathSegment())
+                                );
+                            }
+                        };
+                        break;
+                    default:
+                        span = new ClickableNoLineSpan() {
+                            @Override
+                            public void onClick(View view) {
+                                context.startActivity(SearchResultActivity.getIntent(context, uri.getLastPathSegment()));
+                            }
+                        };
+                        break;
                 }
             } else {
-                return new ClickableSpan() {
+                span = new ClickableNoLineSpan() {
                     @Override
                     public void onClick(View view) {
-                        AppCustomTabsKt.launchChromeCustomTabs(context, link);
+                        AppCustomTabsKt.launchChromeCustomTabs(context, uri);
                     }
                 };
             }
-        };
+            spannableStringBuilder.setSpan(span, link.getStart(), link.getEnd(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
+        }
+        return spannableStringBuilder;
+    }
+
+    public static CharSequence appendLinkAtViaText(Context context, String name, String url) {
+        if (url == null) {
+            return name;
+        } else {
+            SpannableStringBuilder spannableStringBuilder = new SpannableStringBuilder("via:" + name);
+            spannableStringBuilder.setSpan(
+                    new ClickableNoLineSpan() {
+                        @Override
+                        public void onClick(View view) {
+                            AppCustomTabsKt.launchChromeCustomTabs(context, url);
+                        }
+                    }, 4, name.length() + 4, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
+            return spannableStringBuilder;
+        }
     }
 
     public static String convertThumbImageUrl(String baseUrl){
-        return (GlobalApplication.clientType == Type.TWITTER)?
+        return (GlobalApplication.clientType == ClientType.TWITTER)?
                 baseUrl + ":thumb":
                 baseUrl.replace("original", "small");
     }
 
     public static String convertSmallImageUrl(String baseUrl){
-        return (GlobalApplication.clientType == Type.TWITTER)?
+        return (GlobalApplication.clientType == ClientType.TWITTER)?
                 baseUrl + ":small":
                 baseUrl.replace("original", "small");
     }
 
     public static String convertOriginalImageUrl(String baseUrl){
-        return (GlobalApplication.clientType == Type.TWITTER)?
+        return (GlobalApplication.clientType == ClientType.TWITTER)?
                 baseUrl + ":orig":
                 baseUrl;
     }
 
     public static String convertLargeImageUrl(String baseUrl){
-        return (GlobalApplication.clientType == Type.TWITTER)?
+        return (GlobalApplication.clientType == ClientType.TWITTER)?
                 baseUrl + ":large":
                 baseUrl;
     }
@@ -363,9 +268,9 @@ public class TwitterStringUtils {
     }
 
     @StringRes
-    public static int getDidActionStringRes(@Type.ClientTypeInt int type, Action action){
+    public static int getDidActionStringRes(@ClientType.ClientTypeInt int type, Action action){
         switch (type){
-            case Type.TWITTER:
+            case ClientType.TWITTER:
                 switch (action){
                     case LIKE:
                         return R.string.did_like;
@@ -379,7 +284,7 @@ public class TwitterStringUtils {
                         return 0;
                 }
 
-            case Type.MASTODON:
+            case ClientType.MASTODON:
                 switch (action){
                     case LIKE:
                         return R.string.did_favorite;
@@ -399,12 +304,12 @@ public class TwitterStringUtils {
     }
 
     @StringRes
-    public static int getRepeatedByStringRes(@Type.ClientTypeInt int type){
+    public static int getRepeatedByStringRes(@ClientType.ClientTypeInt int type){
         switch (type){
-            case Type.TWITTER:
+            case ClientType.TWITTER:
                 return R.string.retweeted_by;
 
-            case Type.MASTODON:
+            case ClientType.MASTODON:
                 return R.string.boosted_by;
 
             default:
