@@ -19,13 +19,13 @@ package com.github.moko256.twitlatte.widget
 import android.content.Context
 import android.net.Uri
 import android.os.Build
-import android.support.v13.view.inputmethod.EditorInfoCompat
-import android.support.v13.view.inputmethod.InputConnectionCompat
-import android.support.v13.view.inputmethod.InputContentInfoCompat
-import android.support.v7.widget.AppCompatEditText
 import android.util.AttributeSet
 import android.view.inputmethod.EditorInfo
 import android.view.inputmethod.InputConnection
+import androidx.appcompat.widget.AppCompatEditText
+import androidx.core.view.inputmethod.EditorInfoCompat
+import androidx.core.view.inputmethod.InputConnectionCompat
+import androidx.core.view.inputmethod.InputContentInfoCompat
 
 
 /**
@@ -60,7 +60,7 @@ class ImageKeyboardEditText : AppCompatEditText {
             }
             val result = imageAddedListener?.onAdded(inputContentInfo.contentUri)
             if (inputContentInfo.linkUri != null && result == true) {
-                text.append(" " + inputContentInfo.linkUri?.toString())
+                text?.append(" " + inputContentInfo.linkUri?.toString())
             }
             true
         }
